Made VC++ use /W4 instead of /W3 and also to use /permissive- for best warnings and conformance. Renamed copy_image_surface to copy_surface. Moved copy_surface to implementation dependent files. Updated and renamed to_degrees and to_radians. Fixed some warnings in data_source.h. Changed the logic for finding "catch2" in Tests/CMakeLists.txt. Fixed warnings for mask.cpp.

+++ P0267_RefImpl/Tests/mask.cpp
@@ -189,11 +189,11 @@ static image_surface MakeCheckerboardMask(int width, int height)
 {
     image_surface s{format::argb32, width, height};
     s.paint(brush{rgba_color::transparent_black});
-    auto b = brush{rgba_color::white}; 
+    auto b = brush{rgba_color::white};
     for( auto y = 0; y < height; ++y )
         for( auto x = 0; x < width; ++x )
             if( (x + y) % 2 == 0 )
-                s.paint(b, nullopt, nullopt, clip_props{ bounding_box(x, y, 1, 1) });
+                s.paint(b, nullopt, nullopt, clip_props{ bounding_box(static_cast<float>(x), static_cast<float>(y), 1.0f, 1.0f) });
     return s; 
 }
